Added dummy login endpoint as well as proper error recovery and loggin

+++ server/index.js
@@ -1,12 +1,14 @@
 const express = require("express");
 const multer = require("multer");
-const main = multer({ dest: "images/main/" });
 const app = express();
 const cors = require("cors");
 const pool = require("./db");
 const { application } = require("express");
 const crypt = require("bcrypt");
 const fs = require("fs");
+const { restart } = require("nodemon");
+const { DatabaseError } = require("pg-protocol");
+const cookie_parser = require("cookie-parser");
 
 var storage = multer.diskStorage({
   destination: function (req, file, cb) {
@@ -18,6 +20,7 @@ var upload = multer({ storage: storage });
 // middleware
 app.use(cors());
 app.use(express.json());
+app.use(cookie_parser("development"))
 
 // Routes
 
@@ -26,17 +29,70 @@ app.use(express.json());
 app.post("/register", async (req, res) => {
   try {
     const { username, password } = req.body; // Get the username and password from the request body
-    const salt = await crypt.genSalt(); // Generate a random salt for every new password
-    const hash = await crypt.hash(password, salt); // Hash the password with the generated salt
-    await pool.query(
-      "INSERT INTO users (Username, Pass) VALUES($1, $2) RETURNING *",
-      [username, hash]
-    ); // Add the new user to the database
+    // Check if username already exists
+    const exists = await pool.query("SELECT * FROM users WHERE Username = $1", [
+      username,
+    ]);
+    if (username && password) {
+      // User input username & password
+      if (exists.rows[0]) {
+        // Username exists
+        res.send("This Username already exists!");
+      } else {
+        // username does not exist
+        const salt = await crypt.genSalt(); // Generate a random salt for every new password
+        const hash = await crypt.hash(password, salt); // Hash the password with the generated salt
+        await pool.query(
+          "INSERT INTO users (Username, Pass) VALUES($1, $2) RETURNING *",
+          [username, hash]
+        ); // Add the new user to the database
 
-    res.status(201).json("Successfully created user account!");
+        res.status(201).send("Successfully created user account!");
+      }
+    } else {
+      // Either username or password missing
+      res.status(400).send("Please enter Username and Password!");
+    }
   } catch (error) {
     console.error(error.message);
     res.status(500).send("Something went wrong!");
+  }
+});
+
+// POST; Login
+
+app.post("/login", async (req, res) => {
+  try {
+    const { username, password } = req.body; // Get the username and password from the request body
+
+    if (username && password) {
+      // User input username & password
+      const entry = await pool.query(
+        "SELECT pass FROM users WHERE username = $1",
+        [username]
+      ); // Look up if user exists
+
+      if (entry.rows[0]) {
+        // User exists
+        if (await crypt.compare(password, entry.rows[0].pass)) {
+          // Correct password
+          res.status(200).send("Login successful!");
+        } else {
+          // Wrong password
+          res.status(401).send("Wrong password for this username!");
+        }
+      } else {
+        // User does not exist
+        res.status(401).send("This username does not exist!");
+      }
+    } else {
+      // Either username or password missing
+      res.status(400).send("Please enter Username and Password!");
+    }
+  } catch (error) {
+    // An error occurred while trying to process the request
+    console.error(error.message);
+    res.status(500).json("Something went wrong!");
   }
 });
 
@@ -59,14 +115,19 @@ app.post("/recipes", upload.single("main"), async (req, res) => {
 
     const recipeID = newRecipe.rows[0].recipeid;
 
-    await saveCategories(category, recipeID);
-    await saveGroups(groups, recipeID);
-    await saveInstructions(instructions, recipeID);
-    await saveTags(tags, recipeID);
-
-    res.status(201).json(newRecipe.rows[0]);
+    if (
+      (await saveCategories(category, recipeID)) &&
+      (await saveGroups(groups, recipeID)) &&
+      (await saveInstructions(instructions, recipeID)) &&
+      (await saveTags(tags, recipeID))
+    ) {
+      res.status(201).json(newRecipe.rows[0]);
+    } else {
+      res.status(500).send("Something went wrong!");
+    }
   } catch (err) {
     console.error(err.message);
+    res.status(500).send("Something went wrong!");
   }
 });
 
@@ -76,10 +137,11 @@ app.get("/recipes", async (req, res) => {
   try {
     const allRecipes = await pool.query(
       "SELECT * from recipes ORDER BY RecipeID DESC"
-    );
+    ); // Retrieve all recipes from the database
     res.status(200).json(allRecipes.rows);
   } catch (err) {
     console.error(err.message);
+    res.status(500).send("Something went wrong!");
   }
 });
 
@@ -91,22 +153,18 @@ app.get("/recipes/:id", async (req, res) => {
     const recipe = await pool.query(
       "SELECT * from recipes WHERE RecipeID = $1",
       [id]
-    );
+    ); // Get recipe from database
 
-    res.status(200).json(recipe.rows[0]);
+    if (recipe.rows[0]) {
+      // Recipe exists
+      res.status(200).json(recipe.rows[0]);
+    } else {
+      // Recipe does not exist
+      res.status(404).send("This recipe does not exist!");
+    }
   } catch (err) {
     console.error(err.message);
-    res.status(404).json("Unable to find recipe");
-  }
-});
-
-// Some put request
-
-app.put("/recipes/:id", async (req, res) => {
-  try {
-    console.log("Not yet implemented!");
-  } catch (err) {
-    console.error(err.message);
+    res.status(500).json("Something went wrong!");
   }
 });
 
@@ -119,15 +177,20 @@ app.delete("/recipes/:id", async (req, res) => {
       "SELECT * FROM recipes WHERE RecipeID = $1",
       [id]
     ); // Retrieve the recipe from the database
+    if (recipe.rows[0]) {
+      // Recipe exists
+      fs.unlinkSync(String(recipe.rows[0].mainimage)); // Delete the main image related to the database
 
-    fs.unlinkSync(String(recipe.rows[0].mainimage)); // Delete the main image related to the database
+      await pool.query("DELETE FROM recipes WHERE RecipeID = $1", [id]); // Delete the database entry
 
-    await pool.query("DELETE FROM recipes WHERE RecipeID = $1", [id]); // Delete the database entry
-
-    res.status(200).json("Recipe successfully deleted!");
+      res.status(200).send("Recipe deleted successfully!");
+    } else {
+      // Recipe does not exist
+      res.status(404).send("This recipe does not exist");
+    }
   } catch (error) {
     console.error(error.message);
-    res.status(404).json("Unable to find recipe");
+    res.status(500).send("Something went wrong!");
   }
 });
 
@@ -138,24 +201,26 @@ app.listen(5000, () => {
 // Helper function that save instructions to the instruction database
 
 async function saveInstructions(instructions, recipeID) {
-  Object.entries(instructions).forEach(async (instruction) => {
-    try {
+  try {
+    Object.entries(instructions).forEach(async (instruction) => {
       const [step, description] = instruction;
       await pool.query(
         "INSERT INTO recipe_instructions (RecipeID, Step, Instruction, Instruction_Image) VALUES($1, $2, $3, $4) RETURNING *",
         [recipeID, step, description, ""]
       );
-    } catch (error) {
-      console.error(error);
-    }
-  });
+    });
+    return true;
+  } catch (error) {
+    console.error(error);
+    return false;
+  }
 }
 
 // Helper function that checks if recipe tags are not yet part of the database and adds them
 
 async function saveTags(tags, recipeID) {
-  tags.forEach(async (tag) => {
-    try {
+  try {
+    tags.forEach(async (tag) => {
       const checkTag = await pool.query("SELECT * FROM tags WHERE Name = $1", [
         tag,
       ]);
@@ -175,10 +240,12 @@ async function saveTags(tags, recipeID) {
           [recipeID, checkTag.rows[0].tagid]
         );
       }
-    } catch (error) {
-      console.error(error);
-    }
-  });
+    });
+    return true;
+  } catch (error) {
+    console.error(error);
+    return false;
+  }
 }
 
 // Helper function that adds categories related to a recipe to the database
@@ -189,12 +256,18 @@ async function saveCategories(category, recipeID) {
       "SELECT * FROM categories WHERE Name = $1",
       [category]
     );
-    await pool.query(
-      "INSERT INTO recipe_categories (RecipeID, categoryID) VALUES($1, $2) RETURNING *",
-      [recipeID, checkCat.rows[0].categoryid]
-    );
+    if (checkCat.rows[0]) {
+      await pool.query(
+        "INSERT INTO recipe_categories (RecipeID, categoryID) VALUES($1, $2) RETURNING *",
+        [recipeID, checkCat.rows[0].categoryid]
+      );
+    } else {
+      return false;
+    }
+    return true;
   } catch (error) {
     console.error(error);
+    return false;
   }
 }
 
@@ -226,24 +299,28 @@ async function saveIngredients(ingredients, recipeID, groupID) {
         );
       }
     });
+    return true;
   } catch (error) {
     console.error(error);
+    return false;
   }
 }
 
 // Helper function that adds recipe groups to the database
 
 async function saveGroups(groups, recipeID) {
-  Object.entries(groups).forEach(async (group) => {
-    try {
+  try {
+    Object.entries(groups).forEach(async (group) => {
       const [name, values] = group;
       const newGroup = await pool.query(
         "INSERT INTO ingredient_groups (RecipeID, Name) VALUES($1, $2) RETURNING *",
         [recipeID, name]
       );
       await saveIngredients(values, recipeID, newGroup.rows[0].groupid);
-    } catch (error) {
-      console.error(error);
-    }
-  });
+    });
+    return true;
+  } catch (error) {
+    console.error(error);
+    return false;
+  }
 }
